requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
pydantic
python-dotenv
psycopg2-binary
uliweb-alembic
bcrypt
pyjwt
starlette
python-multipart
alembic
requests
cryptography